fix: first vid double of playlist
video playlist now don't add the first video twice anymore during testing this was added because it seemed like the first one would be ignored but now it works again strange behaviour? If this is indeed a behavior condition could be added that checks for it but for now just remove this.